requirements.txt
websocket-client==1.8.0
requests==2.32.3
pandas==2.2.2
pandas-ta==0.3.14b0
yfinance==0.2.40
psycopg2-binary==2.9.9
vaderSentiment==3.3.2

